ScoringEngine/requirements.txt
Flask
flask-login
mechanize
pyaml
PyYAML
SQLAlchemy
pycrypto
psycopg2
enum34
paramiko
bcrypt